get rid of OP_CODE_NAMES_LOOKUP
