+++ buidl/helper.py
@@ -321,48 +321,6 @@ def murmur3(data, seed=0):
     return h1 & 0xFFFFFFFF
 
 
-def number_to_op_code_byte(n):
-    """Returns the OP code for a particular number"""
-    if n < -1 or n > 16:
-        raise ValueError("Not a valid OP code")
-    if n > 0:
-        return bytes([0x50 + n])
-    elif n == 0:
-        return b"\x00"
-    elif n == -1:
-        return b"\x4f"
-
-
-def op_code_to_number(op_code):
-    """Returns the n for a particular OP code"""
-    if op_code not in (
-        0,
-        79,
-        80,
-        81,
-        82,
-        83,
-        84,
-        85,
-        86,
-        87,
-        88,
-        89,
-        90,
-        91,
-        92,
-        93,
-        94,
-        95,
-        96,
-    ):
-        raise ValueError("Not a valid OP code")
-    if op_code == 0:
-        return 0
-    else:
-        return op_code - 80
-
-
 def hmac_sha512(key, msg):
     return hmac.HMAC(key=key, msg=msg, digestmod=hashlib.sha512).digest()
