+++ buidl/op.py
@@ -762,6 +762,58 @@ def op_checksequenceverify(stack, version, sequence):
     return True
 
 
+def number_to_op_code_byte(n):
+    """Returns the OP code for a particular number"""
+    if n < -1 or n > 16:
+        raise ValueError("Not a valid OP code")
+    if n > 0:
+        return bytes([0x50 + n])
+    elif n == 0:
+        return b"\x00"
+    elif n == -1:
+        return b"\x4f"
+
+
+def number_to_op_code(n):
+    if n < -1 or n > 16:
+        raise ValueError("Not a valid OP code")
+
+    if n == 0:
+        return 0
+
+    return n + 80
+
+
+def op_code_to_number(op_code):
+    """Returns the n for a particular OP code"""
+    if op_code not in (
+        0,
+        79,
+        80,
+        81,
+        82,
+        83,
+        84,
+        85,
+        86,
+        87,
+        88,
+        89,
+        90,
+        91,
+        92,
+        93,
+        94,
+        95,
+        96,
+    ):
+        raise ValueError("Not a valid OP code")
+    if op_code == 0:
+        return 0
+    else:
+        return op_code - 80
+
+
 OP_CODE_FUNCTIONS = {
     0: op_0,
     79: op_1negate,
@@ -941,10 +993,3 @@ OP_CODE_NAMES = {
     184: "OP_NOP9",
     185: "OP_NOP10",
 }
-
-OP_CODE_NAMES_LOOKUP = {}
-for k in OP_CODE_NAMES:
-    OP_CODE_NAMES_LOOKUP[OP_CODE_NAMES[k]] = k
-
-# Sanity check:
-assert len(OP_CODE_NAMES) == len(OP_CODE_NAMES_LOOKUP)
